requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aioredis==2.0.1
    # via fastray (pyproject.toml)
aiosqlite==0.21.0
    # via tortoise-orm
amqp==5.3.1
    # via kombu
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via starlette
async-timeout==5.0.1
    # via aioredis
billiard==4.2.1
    # via celery
celery==5.4.0
    # via fastray (pyproject.toml)
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   uvicorn
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
colorama==0.4.6
    # via click
fastapi==0.115.11
    # via fastray (pyproject.toml)
greenlet==3.1.1
    # via sqlalchemy
h11==0.14.0
    # via uvicorn
idna==3.10
    # via
    #   anyio
    #   requests
iso8601==2.1.0
    # via tortoise-orm
jinja2==3.1.6
    # via fastray (pyproject.toml)
kombu==5.5.0
    # via celery
markupsafe==3.0.2
    # via jinja2
numpy==2.2.4
    # via shapely
passlib==1.7.4
    # via fastray (pyproject.toml)
prompt-toolkit==3.0.50
    # via click-repl
pydantic==2.10.6
    # via
    #   fastapi
    #   pydantic-settings
    #   sqlmodel
pydantic-core==2.27.2
    # via pydantic
pydantic-settings==2.8.1
    # via fastray (pyproject.toml)
pyjwt==2.10.1
    # via fastray (pyproject.toml)
pymysql==1.1.1
    # via fastray (pyproject.toml)
pypika-tortoise==0.5.0
    # via tortoise-orm
python-dateutil==2.9.0.post0
    # via celery
python-dotenv==1.0.1
    # via
    #   fastray (pyproject.toml)
    #   pydantic-settings
python-multipart==0.0.20
    # via fastray (pyproject.toml)
pytz==2025.1
    # via tortoise-orm
requests==2.32.3
    # via fastray (pyproject.toml)
shapely==2.0.7
    # via fastray (pyproject.toml)
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.39
    # via sqlmodel
sqlmodel==0.0.24
    # via fastray (pyproject.toml)
starlette==0.46.1
    # via fastapi
tortoise-orm==0.24.2
    # via fastray (pyproject.toml)
typing-extensions==4.12.2
    # via
    #   aioredis
    #   aiosqlite
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
tzdata==2025.1
    # via
    #   celery
    #   kombu
urllib3==2.3.0
    # via requests
uvicorn==0.34.0
    # via fastray (pyproject.toml)
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
